Modularized the code necessary for drawing SVG Elements
